+++ docs/source/helpers.rst
@@ -12,3 +12,4 @@ libcst.helpers
 
 .. autofunction:: libcst.helpers.module.insert_header_comments
 .. autofunction:: libcst.helpers.expression.get_full_name_for_node
+.. autofunction:: libcst.ensure_type
